refactor cache importer into separate package and merge with master
